+++ tests/app/main/forms/test_strip_whitespace.py
